+++ client/src/app/app.module.ts
@@ -19,12 +19,12 @@ import { DeviceComponent } from "./views/devices/device/device.component";
 import { DatasetComponent } from "./views/datasets/dataset/dataset.component";
 import { MatTabsModule } from "@angular/material/tabs";
 import { LineChartNgxComponent } from "./components/line-chart-ngx/line-chart-ngx.component";
-import { NewDeviceDialogComponent } from "./dialogs/new-device/new-device.component";
+import { NewDeviceDialogComponent } from "./components/dialogs/new-device/new-device.component";
 import { MatDialogModule } from "@angular/material/dialog";
 import { MatCardModule } from "@angular/material/card";
 import { MatInputModule } from "@angular/material/input";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
-import { NewDatasetDialogComponent } from "./dialogs/new-dataset/new-dataset.component";
+import { NewDatasetDialogComponent } from "./components/dialogs/new-dataset/new-dataset.component";
 import { MatSelectModule } from "@angular/material/select";
 import { NgxChartsModule } from "@swimlane/ngx-charts";
 import { MatProgressSpinnerModule } from "@angular/material/progress-spinner";
@@ -34,7 +34,7 @@ import { EventsComponent } from "./views/events/events.component";
 import { MatTableModule } from "@angular/material/table";
 import { MatSortModule } from "@angular/material/sort";
 import { HttpErrorInterceptor } from "./services/httperrorinterceptor.service";
-import { ErrorHandlingDialogComponent } from "./dialogs/error-handling-dialog/error-handling-dialog.component";
+import { ErrorHandlingDialogComponent } from "./components/dialogs/error-handling-dialog/error-handling-dialog.component";
 import { FlexLayoutModule } from "@angular/flex-layout";
 import { DatasetDetailsComponent } from "./components/dataset-details/dataset-details.component";
