+++ client/src/app/components/dialogs/new-device/new-device.component.html
